Added a method retrive a formated API response with extra information
I decided to add this instance method because I didn't want to store duplicate data in the database for the actual shortlink url, so this allows me just calculate the link string and include it in the response. Improving API usablity without adding extra data to the database. I could've also used virtuals, but including virtuals in the toJSON output also seemed to add a lot of other internal properties that wouldn't be useful to the end user.