Added creation error messages
I might try to update this later on to better take advantage of the backend error responses, but first I would want to improve the backend responses because now that I'm actually working on the frontend I feel like they would need to be made a little more user friendly for them to be more useful to the user. Will see if I every get around to that concidering this is a side project that I've already spent way too much time on. I'm taking this way too seriously, lol.